command リファクタ & 重複した返答メッセージ消した
