+++ src/Channel/cmd/QUIT.cpp
@@ -1,6 +1,15 @@
 #include "Client.hpp"
 #include "ChannelManager.hpp"
 
+
+void Channel::quit(const Client &target, const std::string &quit_msg)
+{
+    assert(is_member(target));
+    members.erase(target);
+    if (is_operator(target))
+        operators.erase(target);
+}
+
 //QUIT [:msg] : msgは省略可、サーバーから抜け、全てのクライアントにメッセージを送る
 void ChannelManager::quit_all_channel(Client &client, const Command& cmd)
 {
